Refactoring of Season Card
