+++ package.json
@@ -91,7 +91,7 @@
     "prettier": "3.3.2",
     "prettier-plugin-tailwindcss": "^0.6.5",
     "storybook": "^8.1.11",
-    "typescript": "^5.5.3",
+    "typescript": "^5.6.0-dev.20240724",
     "typescript-eslint": "^7.15.0"
   },
   "prettier": {
