+++ src/components/UnofficialLabel.tsx
@@ -1,0 +1,10 @@
+export const UnofficialLabel = () => {
+  return (
+    <span
+      className="absolute left-4 top-2 font-mono text-xs font-semibold tracking-wider text-yellow-400 opacity-75 md:text-sm"
+      title="This is not an official game or cycle, which means it is not actively supported by the publishers."
+    >
+      UNOFFICIAL*
+    </span>
+  );
+};
